Refactor project to deal with chips as BigDecimal

+++ src/main/java/com/gui/jhand/action/Action.java
@@ -4,6 +4,8 @@ import lombok.AllArgsConstructor;
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
 
+import java.math.BigDecimal;
+
 @Getter
 @EqualsAndHashCode
 @AllArgsConstructor(staticName = "of")
@@ -17,7 +19,7 @@ public class Action {
 
 	private String mainInformation;
 
-	private double amount;
+	private BigDecimal amount;
 
 	private String rawLine;
